♻️ Outsource CartContextProvider
