+++ components/ItemDetails.tsx
@@ -1,15 +1,17 @@
-import React from "react";
+import React, { useContext } from "react";
+import { CartContext } from "../contexts/CartContext";
 import { Item } from "../utils/types";
 
 type ItemDetailsProps = {
   item: Item;
-  onAdd(): void;
 };
-function ItemDetails({ item, onAdd }: ItemDetailsProps) {
+function ItemDetails({ item }: ItemDetailsProps) {
+  const { onItemAdd } = useContext(CartContext);
+
   return (
     <>
       {item.name} <b>${item.price}</b>
-      <button onClick={onAdd}>Add to cart</button>
+      <button onClick={() => onItemAdd(item)}>Add to cart</button>
     </>
   );
 }
